+++ src/slskd/SourceFeeds/ISpotifyConnectionService.cs
@@ -20,4 +20,3 @@ public interface ISpotifyConnectionService
 
     void Disconnect();
 }
-